Changed how the history works
Changed how the history works to completely remove duplicated entries in the future. Instead of creating a new entry everytime while fetching, the last entry is being checked for differences. If the new fetch is different from the last history, a new entry is created, otherwise `updated_at` is touched, to signal that the entry was created earlier, but is still upto date.